requirements.txt
python-telegram-bot==21.6
groq==0.11.0
supabase==2.9.1
httpx==0.27.2
Pillow==10.4.0
Flask==3.0.3
Flask-Login==0.6.3
python-dotenv==1.0.1
aiofiles==24.1.0
apscheduler==3.10.4
gunicorn==23.0.0
